+++ tf/about/variants.py
@@ -1,0 +1,3 @@
+"""
+.. include:: ../docs/about/variants.md
+"""